requirements.txt
python-dotenv>=1.0.0
netifaces>=0.11.0
# 开发和测试依赖
# pytest>=8.0.0
# pytest-mock>=3.0.0
# pytest-cov>=5.0.0
